+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # CHANGELOG
 
 ## 1.4.1
+### Bug fix
  - Allow init import and product import having numeric identifier
+
+### BC break
+ - Inject `pim_catalog.repository.attribute` inside `pim_excel_connector.reader.spreadsheet` with `setAttributeRepository` method
 
 ## 1.4.0
  - Made compatible with PIM CE 1.3.x